+++ src/Definition.php
@@ -4,10 +4,14 @@ declare(strict_types=1);
 namespace Capsule\Di;
 
 use Capsule\Di\Lazy\Lazy;
+use Throwable;
+use Capsule\Di\Exception;
 
 abstract class Definition extends Lazy
 {
     protected mixed /* callable */ $factory = null;
+
+    protected bool $isInstantiable = false;
 
     public function __invoke(Container $container) : mixed
     {
@@ -20,5 +24,30 @@ abstract class Definition extends Lazy
         return $this;
     }
 
-    abstract public function new(Container $container) : object;
+    public function isInstantiable(Container $container) : bool
+    {
+        if ($this->factory !== null) {
+            return true;
+        }
+
+        if ($this->class !== null) {
+            return $container->has($this->class);
+        }
+
+        return $this->isInstantiable;
+    }
+
+    public function new(Container $container) : object
+    {
+        try {
+            return $this->instantiate($container);
+        } catch (Throwable $e) {
+            throw new Exception\NotInstantiated(
+                "Could not instantiate {$this->id}",
+                previous: $e
+            );
+        }
+    }
+
+    abstract protected function instantiate(Container $container) : object;
 }