+++ src/Exception/NotInstantiated.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Capsule\Di\Exception;
+
+class NotInstantiated extends Exception
+{
+}